+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased]
+## [Unreleased]
 
 ### Released TBA
 
-[unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.79.0-sumo-0...main -->
+### Changed
+
+- chore(sumologicexporter): remove deprecation error messages [#1167]
+
+[#1167]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1167
+
+[unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.79.0-sumo-0...main
 
 ## [v0.79.0-sumo-0]