+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.80.0-sumo-0]
 
-### Released TBA
+### Released 2023-06-30
 
 ### Changed
 
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#1174]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1174
 [#1180]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1180
 
-[unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.79.0-sumo-0...main
+[v0.80.0-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.79.0-sumo-0...v0.78.0-sumo-0
 
 ## [v0.79.0-sumo-0]